requirements.txt
pyserial==3.5
gpiozero==2.0.1
lgpio==0.2.2.0
